# apps/backend/src/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Logistgo Pro API",
    "version": "1.0.0",
    "description": "REST API для Logistic Pro. Полнофункциональная система управления логистическими компаниями с поддержкой:\n\n• **Аутентификация и авторизация** - регистрация, вход, JWT токены\n• **Управление компаниями** - создание, просмотр, типы компаний\n• **Управление пользователями** - добавление, приглашение, изменение ролей\n• **Система ролей** - Владелец, Администратор, Пользователь\n• **Публичные API** - поиск автомобилей, поиск мест, построение маршрутов\n• **Поиск попутных грузов** - поиск грузов по маршруту пользователя\n• **Интеграция с картами** - Nominatim, OSRM для геокодирования и маршрутизации\n• **Интерактивные карты** - создание грузов с визуализацией маршрута и промежуточными точками\n\nТокены формата JWT передавать в заголовке `Authorization: Bearer <token>`.\nПубличные эндпоинты (поиск автомобилей, мест, маршрутов) не требуют авторизации.\n\n**Базовый URL:** `http://0.0.0.0:5555/api`"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation error",
            "description": "Краткое описание ошибки."
          },
          "error": {
            "type": "string",
            "description": "Сообщение об ошибке (если есть)."
          },
          "errors": {
            "type": "object",
            "description": "Детализация ошибок валидации."
          }
        },
        "description": "Стандартное представление ошибки."
      },
...
      "CreateCargoRequest": {
        "type": "object",
        "required": [
          "id_company",
          "departure_point",
          "arrival_point",
          "id_car_type",
          "id_tip_zagryzki",
          "tonn",
          "m3",
          "payment",
          "date_start",
          "date_end"
        ],
        "properties": {
          "id_company": {
            "type": "string",
            "format": "uuid"
          },
          "departure_point": {
            "type": "string",
            "example": "Минск"
          },
          "arrival_point": {
            "type": "string",
            "example": "Гродно"
          },
          "id_car_type": {
            "type": "integer",
            "example": 1
          },
          "id_tip_zagryzki": {
            "type": "integer",
            "example": 2
          },
          "opisanie": {
            "type": "string",
            "example": "Нужна перевозка стройматериалов"
          },
          "tonn": {
            "type": "number",
            "example": 10
          },
          "m3": {
            "type": "number",
            "example": 20
          },
          "price": {
            "type": "string",
            "example": "1000 BYN"
          },
          "payment": {
            "type": "string",
            "enum": [
              "Наличный",
              "Безналичный",
              "Карта",
              "Перевод"
            ]
          },
          "date_start": {
            "type": "string",
            "format": "date-time"
          },
          "date_end": {
            "type": "string",
            "format": "date-time"
          },
          "irrelevant": {
            "type": "integer",
            "example": 0
          },
          "departure_place_id": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "arrival_place_id": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "status": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "UpdateCargoRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateCargoRequest"
          }
        ]
      },
      "CreateCarRequest": {
        "type": "object",
        "required": [
          "id_company",
          "title",
          "id_car_type",
          "id_tip_zagryzki",
          "tonn_min",
          "tonn_max",
          "m3_min",
          "m3_max"
        ],
        "properties": {
          "id_company": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Тент 20т"
          },
          "id_car_type": {
            "type": "integer",
            "example": 1
          },
          "id_tip_zagryzki": {
            "type": "integer",
            "example": 2
          },
          "tonn_min": {
            "type": "number",
            "example": 10
          },
          "tonn_max": {
            "type": "number",
            "example": 20
          },
          "m3_min": {
            "type": "number",
            "example": 30
          },
          "m3_max": {
            "type": "number",
            "example": 60
          },
          "price": {
            "type": "number",
            "example": 1500
          },
          "subscription": {
            "type": "boolean",
            "example": true
          },
          "search": {
            "type": "boolean",
            "example": false
          },
          "places": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateCarRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateCarRequest"
          }
        ]
      },
...
      "CreateRouteRequest": {
        "type": "object",
        "required": [
          "departure_point",
          "arrival_point"
        ],
        "properties": {
          "id_cars": {
            "type": "integer",
            "example": 1
          },
          "departure_point": {
            "type": "string",
            "example": "Минск"
          },
          "arrival_point": {
            "type": "string",
            "example": "Брест"
          },
          "places_departure": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "places_arrival": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "date_start": {
            "type": "string",
            "format": "date-time"
          },
          "opisanie": {
            "type": "string",
            "example": "Еду пустой, ищу загрузку по пути"
          }
        }
      },
      "UpdateRouteRequest": {
        "type": "object",
        "properties": {
          "departure_point": {
            "type": "string",
            "example": "Минск"
          },
          "arrival_point": {
            "type": "string",
            "example": "Брест"
          },
          "places_departure": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "places_arrival": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lon": {
                  "type": "number"
                }
              }
            }
          },
          "date_start": {
            "type": "string",
            "format": "date-time"
          },
          "opisanie": {
            "type": "string",
            "example": "Еду пустой, ищу загрузку по пути"
          }
        },
        "description": "Тело запроса для обновления маршрута (без идентификаторов). Идентификатор маршрута передаётся в path как id_routes."
      },
...
      "OSRMBuildRouteRequest": {
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "start": {
            "type": "object",
            "required": [
              "lat",
              "lon"
            ],
            "properties": {
              "lat": {
                "type": "number",
                "example": 53.9006
              },
              "lon": {
                "type": "number",
                "example": 27.559
              }
            }
          },
          "end": {
            "type": "object",
            "required": [
              "lat",
              "lon"
            ],
            "properties": {
              "lat": {
                "type": "number",
                "example": 52.4242
              },
              "lon": {
                "type": "number",
                "example": 31.0143
              }
            }
          },
          "avoidMotorwayToll": {
            "type": "boolean",
            "example": false,
            "description": "Избегать платных автомагистралей"
          },
          "preferShortest": {
            "type": "boolean",
            "example": false,
            "description": "Предпочитать кратчайший путь"
          }
        },
        "description": "Параметры построения маршрута через OSRM"
      },
      "OSRMBuildRouteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "distance": {
                "type": "number",
                "example": 300000,
                "description": "Расстояние в метрах"
              },
              "duration": {
                "type": "number",
                "example": 10800,
                "description": "Время в секундах"
              },
              "geometry": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "LineString"
                  },
                  "coordinates": {
                    "type": "array",
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      },
                      "minItems": 2,
                      "maxItems": 2
                    },
                    "example": [
                      [
                        27.559,
                        53.9006
                      ],
                      [
                        31.0143,
                        52.4242
                      ]
                    ]
                  }
                }
              },
              "nodes": {
                "type": "array",
                "items": {
                  "type": "integer"
                },
                "example": [
                  12345,
                  12346,
                  12347
                ],
                "description": "Массив узлов маршрута (опционально)"
              }
            }
          }
        },
        "description": "Результат построения маршрута через OSRM"
      },
...
    }
  },
  "paths": {
...
    "/api/cargo": {
      "post": {
        "summary": "Создать груз",
        "description": "Создание карточки груза. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCargoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Груз создан"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          }
        }
      }
    },
    "/api/cargo/{id}": {
      "patch": {
        "summary": "Обновить груз",
        "description": "Редактирование карточки груза по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCargoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Груз обновлён"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "delete": {
        "summary": "Удалить груз",
        "description": "Удаление карточки груза по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      }
    },
...
    "/api/cars": {
      "post": {
        "summary": "Создать автомобиль",
        "description": "Создание карточки автомобиля. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Автомобиль создан"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          }
        }
      }
    },
    "/api/cars/{id}": {
      "get": {
        "summary": "Получить автомобиль по ID",
        "description": "Возвращает данные автомобиля по его ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID автомобиля"
          }
        ],
        "responses": {
          "200": {
            "description": "Автомобиль найден"
          },
          "400": {
            "description": "Некорректный идентификатор"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "patch": {
        "summary": "Обновить автомобиль",
        "description": "Редактирование карточки автомобиля по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Автомобиль обновлён"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "delete": {
        "summary": "Удалить автомобиль",
        "description": "Удаление карточки автомобиля по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      }
    },
...
    "/api/routes": {
      "post": {
        "summary": "Создать маршрут",
        "description": "Создание маршрута перевозки. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Маршрут создан"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          }
        }
      }
    },
    "/api/routes/{id_routes}": {
      "patch": {
        "summary": "Обновить маршрут",
        "description": "Редактирование маршрута по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id_routes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID маршрута (id_routes)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Маршрут обновлён"
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "delete": {
        "summary": "Удалить маршрут",
        "description": "Удаление маршрута по ID. Требуется авторизация.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id_routes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID маршрута (id_routes)"
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "401": {
            "description": "Требуется авторизация"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      }
    },
...
    "/api/osrm/route": {
      "post": {
        "summary": "Построение маршрута через OSRM",
        "description": "Построение маршрута между двумя точками через сервис OSRM. Публичный эндпоинт, не требует авторизации.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OSRMBuildRouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Маршрут успешно построен через OSRM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OSRMBuildRouteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка сервиса OSRM",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social-cargo": {
      "get": {
        "summary": "Получить список грузов из социальных сетей",
        "description": "Публичный эндпоинт для получения списка грузов из социальных сетей. Поддерживает фильтрацию и пагинацию. Если пользователь авторизован, грузы, отмеченные им как неактуальные, будут скрыты.",
        "tags": [
          "Social Cargo"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Количество грузов (по умолчанию 20, максимум 250)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 250,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Смещение для пагинации",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "contains",
            "in": "query",
            "description": "Фильтр: объявление содержит (можно указать несколько слов через запятую)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notContains",
            "in": "query",
            "description": "Фильтр: объявление НЕ содержит (можно указать несколько слов через запятую)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список грузов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social-cargo/{id}/irrelevant": {
      "post": {
        "summary": "Отметить груз как неактуальный",
        "description": "Отмечает груз как неактуальный для текущего пользователя. Создает запись в таблице report и увеличивает счетчик irrelevant в cargo_messendger. Если груз получает 3 или более отметок, он скрывается для всех пользователей.",
        "tags": [
          "Social Cargo"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID груза",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Груз успешно отмечен как неактуальный",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id_cargo": {
                          "type": "integer"
                        },
                        "irrelevant": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Груз уже отмечен как неактуальный или некорректный ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}